Refactor meal models structure
