+++ mbd-web/src/lib/api/contracts.ts
@@ -1,0 +1,11 @@
+export interface ApiFood {
+  food_id?: string;
+  name: string;
+  thumbnail?: string;
+}
+
+export interface ApiMeal {
+  mealType: string;
+  dateTime: string;
+  foods: ApiFood[];
+}
