+++ mbd-web/src/lib/features/meals/models.ts
@@ -1,0 +1,18 @@
+export interface Food {
+  id: string;
+  name: string;
+  thumbnail?: string;
+}
+
+export interface Meal {
+  id: string;
+  mealType: string;
+  dateTime: string;
+  foods: Food[];
+}
+
+export interface MealsState {
+  meals: Meal[];
+  loading: boolean;
+  daysLoaded: number;
+}
